+++ src/TextUtils/Headings.hs
@@ -22,6 +22,8 @@ import Data.List.Split (splitOn)
 import qualified Data.Map as Map
 
 import Config
+import Data.Maybe (listToMaybe, catMaybes, isJust)
+import Data.Foldable (traverse_)
 
 
 -- | An ASCII art character is represented by breaking up each line into an element of a list.
@@ -35,31 +37,90 @@ type AsciiFont = Map.Map Char AsciiChar
 
 -- | Parse a Burrow ASCII art font file.
 --
+-- The character width and height is determined from the first entry.
+--
 -- See the project's README for the font spec.
 parseFont :: FilePath -> IO AsciiFont
 parseFont path = do
   fontFileContents <- readFile path
-  pure $ Map.fromList $ map charLegendsToPair (splitIntoCharLegends fontFileContents)
+  let
+    charPairs = map charLegendsToPair (splitIntoCharLegends fontFileContents)
+    (width, height) = case charPairs of
+      [] -> error "No characters found in font file."
+      (_, firstChar):_ ->
+        (length (head firstChar), length firstChar)  -- FIXME: use of head here is unsafe
+    (validatedCharPairs, warnings) = unzip $ (flip map) charPairs $ \(charTitle, character) ->
+      let
+        (validatedChar, warnings') = validateCharacter width height charTitle character
+        validatedPair = (charTitle, validatedChar)
+      in
+        (validatedPair, warnings')
+  -- display any warnings for this font
+  if any isJust warnings
+    then do
+      putStrLn $ "Warnings encountered while parsing font file: " ++ path ++ "."
+      traverse_ putStrLn (catMaybes warnings)
+    else
+      pure ()
+  pure $ Map.fromList validatedCharPairs
  where
   splitIntoCharLegends = splitOn "\n\n"
   charLegendsToPair charLegend =
     case lines charLegend of
-      _:[] -> error "Char legend misformatted."
+      _:[] -> error "Font file is not divided into sections by blank lines."
       x:xs ->
         case x of
           c:[] -> (c, xs)
-          _ -> error "Legend in char legend is not just a single character."
+          _ ->
+            error $ "Legend in char legend is not just a single character, ensure if first entry there are no newlines " ++
+            "preceding, double check for whitespace, ensure clear double newlines between the entries:\n\n" ++ unlines xs
       _ -> error "Char legend misformatted."
 
-  -- i can't think of any useful flags for the font!
-  -- Parse the meta/header where flags are defined and return the rest.
-  --parseMeta
+-- | Determine if a character has any errors, attempt to automatically fix them and return
+-- any warnings. May possibly raise an error.
+validateCharacter
+  :: Int
+  -- ^ The width the character is supposed to be.
+  -> Int
+  -- ^ The height the character is supposed to be.
+  -> Char
+  -- ^ The character title. The key for the pair. Used for errors.
+  -> [String]
+  -- ^ The to-be AsciiChar to check.
+  -> (AsciiChar, Maybe String)
+  -- ^ The character after any automatic fixes.
+validateCharacter width height characterTitle asciiChar = do
+  let
+    -- height
+    (heightValidated, heightWarning) = case heightCheck asciiChar of
+      Nothing -> (asciiChar, Nothing)
+      Just a -> (heightFix asciiChar, Just a)
+    -- width
+    (widthValidated, widthWarning) = case widthCheck heightValidated of
+      Nothing -> (heightValidated, Nothing)
+      Just a -> (widthFix heightValidated, Just a)
+    -- combine warnings
+    warnings = catMaybes [heightWarning, widthWarning]
+  -- combine warnings and give the final result
+  (widthValidated, if null warnings then Nothing else Just $ unlines warnings)
+ where
+  heightCheck char = if length char /= height then Just ("Expected height (" ++ show height ++ ") is not consistent for \"" ++ [characterTitle] ++ "\"") else Nothing
+  heightFix char
+    | length char > height = take height char
+    | length char < height = take height $ char ++ replicate (height - length char) (replicate width ' ')
+    | otherwise = char
+
+  widthCheck char = if any (\x -> length x /= width) char then Just ("Expected width (" ++ show width ++ ") is inconsistent for \"" ++ [characterTitle] ++ "\"") else Nothing
+  widthFix char =
+    let lineFix line = if length line > width then take width line else line ++ replicate (width - length line) ' '
+    in map lineFix char
 
 
 -- | Specifies which AsciiFont is related to which heading level.
 type HeadingLevelFontMap = Map.Map Int AsciiFont
 
 
+-- FIXME: if defined more than once, will load in the same fonts over and over!
 -- | Get all the fonts loaded, mapped to a specific heading level,
 -- according to the configuration file.
 getAsciiFonts :: IO HeadingLevelFontMap
@@ -68,7 +129,6 @@ getAsciiFonts = do
   let func level = getConfigValue configParser "fonts" ("h" ++ (show level)) >>= parseFont >>= pure . (,) level
   result <- traverse func [1..6]
   pure $ Map.fromList result
-
 
 -- SHOULD DOCUMENT THE BEHAVIOR OF LOOKING UP. if failure to look up the requested case
 -- then get the opposite.
@@ -90,11 +150,9 @@ fontLookup c f =
 -- | The font must be of a consistent height. We sample the first character
 -- we come across.
 getCharacterHeight :: AsciiFont -> Int
-getCharacterHeight font =
-  -- FIXME/TODO: just use a error exception handler
-  if Map.null font
-    then error "While trying to get height: no characters defined."
-    else length $ head $ Map.elems font
+getCharacterHeight font
+  | Map.null font = error "While trying to get width: no characters defined."
+  | otherwise = maybe (error "While trying to get height: character list is empty.") length (listToMaybe $ Map.elems font)
 
 
 -- | The main function: ...